contribute domain-connect-template-linter tool
This tool tries to make it easy to check and see if a template is valid.
The tool has Cloudflare speficic -cloudflare option.  This option is meant
for service providers who want to onboard to Cloudflare.  Other DNS
providers are welcome to add their own hints about settings.

The basic validation attempts to follow the Domain Connect Spec as close as
possible.  When validation issues happen output is provided to stderr.

Finally there is -pretty option that will unify template json format.  This
is technically irrelevant, but it is nice for humans to have some regularity
in the template files.

Reference: https://github.com/Domain-Connect/spec/blob/master/Domain%20Connect%20Spec%20Draft.adoc#52-template-definition
Signed-off-by: Sami Kerola <kerolasa@cloudflare.com>