Ensure dc-template-linter is checking template logo reachability
Logo reachability was considered optional to start with my thinking was that
logos are not really that important, and the linter will become considerably
slower when testing logos.  Additionally the test makes linter to require
network reachability, that may or may not be always guaranteed.  In short,
it's good to have this as optional so that where networking does not work
majority of the validation will still work.

Reference: 0a1a0cae068118e848d916af227368aa8a840daf
Signed-off-by: Sami Kerola <kerolasa@iki.fi>